Add support of accepting output file name in aadl_xml
The original implementation of the aadl_xml backend takes the
command line option "-o" as the output directory and lets the
XML generator determines the output XML file name based on the
given AADL model's name.

This commit adds support of accepting the "-o" option as either
an output directory or an output file name. In the case that a
directroy is specified or the option "-o" is not assigned, the
output file name is determined by the XML generator as it is
originally implementated. Otherwise, the output XML is stored
in the file with the given file anme.

This change also enables testing the aadl_xml backend under the
current test environment.

Change-Id: I3dd4a7e12784ae5fe81a9ae59271d68913ad6109